# repo/backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Philatelic Museum Operations Suite API
  version: 1.0.0
  description: |
    Public HTTP API for museum operations.

    Authentication model:
    - Session cookie `museum_sid` (set on `POST /auth/login`)
    - `X-CSRF-Token` required on mutating requests (`POST|PUT|PATCH|DELETE`) except `POST /auth/login`
    - `X-Step-Up-Token` required on sensitive actions after `POST /auth/step-up`

    Step-up actions:
    - `GRAPH_PUBLISH`
    - `ROUTE_RULE_CHANGE`
    - `JOB_APPROVE`
    - `JOB_APPEAL_DECIDE`
    - `EXPORT_CREATE`
    - `ADMIN_CONFIG_UPDATE`
servers:
  - url: http://localhost:8080/api/v1
...
security:
  - SessionCookie: []
paths:
...
  /routes:
    post:
      tags: [Venues & Routes]
      summary: Create route definition
      description: Requires `VENUE_MANAGE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      responses:
        '201': { description: Route created }
...
  /routes/{routeId}:
    patch:
      tags: [Venues & Routes]
      summary: Update route rules
      description: Requires `ROUTE_RULE_CHANGE` and step-up action `ROUTE_RULE_CHANGE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
          StepUpHeader: []
      parameters:
        - name: routeId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Route updated }
...
  /inbox/messages:
    get:
      tags: [Inbox]
      summary: List inbox messages for authenticated user
      description: Requires authenticated session; results are recipient-scoped.
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 50 }
        - name: filter[unread]
          in: query
          schema: { type: string, enum: ['true', 'false'] }
        - name: filter[type]
          in: query
          schema: { type: string }
      responses:
        '200': { description: Message list }
  /inbox/messages/{messageId}/read:
    post:
      tags: [Inbox]
      summary: Mark inbox message as read
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Read timestamp updated }
        '404': { $ref: '#/components/responses/NotFound' }
  /inbox/messages/{messageId}/print:
    post:
      tags: [Inbox]
      summary: Retrieve printable payload for message
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Printable payload }
        '404': { $ref: '#/components/responses/NotFound' }
  /jobs:
    post:
      tags: [Jobs]
      summary: Create job draft
      description: Requires `JOB_EDIT` (Administrator, Employer).
      security:
        - SessionCookie: []
          CsrfHeader: []
      responses:
        '201': { description: Draft created }
  /jobs/{jobId}:
    patch:
      tags: [Jobs]
      summary: Edit job draft content
      description: Requires `JOB_EDIT` plus employer ownership for non-admin users.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Job updated }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /jobs/{jobId}/submit:
    post:
      tags: [Jobs]
      summary: Submit job for approval
      description: Requires `JOB_EDIT` plus employer ownership for non-admin users.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Job moved to pending approval }
  /jobs/{jobId}/approve:
    post:
      tags: [Jobs]
      summary: Approve job and publish
      description: Requires `JOB_APPROVE` and step-up action `JOB_APPROVE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
          StepUpHeader: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            example:
              comment: Approved in workflow
      responses:
        '200':
          description: Approved
          content:
            application/json:
              example:
                data:
                  jobId: 65f0f7d1d5f5f8612f06d001
                  state: PUBLISHED
  /jobs/{jobId}/reject:
    post:
      tags: [Jobs]
      summary: Reject job back to draft
      description: Requires `JOB_APPROVE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Rejected }
  /jobs/{jobId}/takedown:
    post:
      tags: [Jobs]
      summary: Takedown published job
      description: Requires `JOB_APPROVE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Takedown recorded }
  /jobs/{jobId}/appeals:
    post:
      tags: [Jobs]
      summary: Submit appeal for takedown
      description: Requires `JOB_EDIT` plus employer ownership for non-admin users.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      responses:
        '201': { description: Appeal created }
  /jobs/{jobId}/appeals/{appealId}/decide:
    post:
      tags: [Jobs]
      summary: Decide appeal
      description: Requires `JOB_APPROVE` and step-up action `JOB_APPEAL_DECIDE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
          StepUpHeader: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
        - name: appealId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            example:
              decision: APPROVE
              comment: Reinstated after review
      responses:
        '200': { description: Appeal decision recorded }
  /jobs/{jobId}/history:
    get:
      tags: [Jobs]
      summary: Job version/workflow history
      description: Requires `JOB_READ`; non-admin employers are owner-scoped.
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Job history payload }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: museum_sid
      description: Session cookie set by `POST /auth/login`.
    CsrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required on mutating requests except `POST /auth/login`.
    StepUpHeader:
      type: apiKey
      in: header
      name: X-Step-Up-Token
      description: One-time action-bound step-up token for sensitive operations.
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              oneOf:
                - type: array
                  items: { type: object }
                - type: object
            requestId: { type: string }
  responses:
...
    Forbidden:
      description: Insufficient role permission or object-level access denied
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: State conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...